refactor(runtime): move prepare-step builders and recorders into AiSdkCompaction
Move the prepare-step builder family (prepareContextBudgetPolicy,
buildActiveToolResultPrunePrepareStep, buildSemanticCompactPrepareStep,
buildActiveFullCompactPrepareStep), the semantic/active-full-compact
recorders, the context-budget policy prep, the message-signature family,
and the compose/projectAccepted/hasPrune helpers from AiSdkBackend into the
AiSdkCompaction collaborator.

The collaborator gains modelAdapter and computeCostUsd dependencies, and
buildSemanticCompactPrepareStep takes an abortSignal threaded from send().
Helpers still used by the co-located mid-turn family
(modelMessageSignature, projectAcceptedActiveFullCompactMessages,
ActiveFullCompactPrepareStepProjection) plus composeActiveCompactionPrepareStep
and hasActiveToolResultPruneDiagnosticPatch are exported back to AiSdkBackend.

Behavior-neutral pure move: 2073 pass / 0 fail / 7 skipped.

Part of #1084
