+++ packages/runtime/src/ai-sdk-compaction.ts
@@ -14,7 +14,7 @@ import type {
   BackendCompactHistoryInput,
   BackendCompactHistoryResult,
 } from '@maka/core/backend-types';
-import type { ContextBudgetDiagnostic } from '@maka/core/usage-stats/types';
+import type { ContextBudgetDiagnostic, LlmCallRecord } from '@maka/core/usage-stats/types';
 
 import type { AiSdkBackendInput } from './ai-sdk-backend.js';
 import {
@@ -22,14 +22,18 @@ import {
   historyCompactBlockToCompactionBoundary,
 } from './compaction-boundary.js';
 import {
+  ARCHIVED_TOOL_RESULT_REWRITE_VERSION,
   applyRuntimeEventContextBudget,
   buildContextBudgetDiagnosticShell,
   estimateRuntimeEventsTokens,
   mergeContextBudgetDiagnostic,
+  mergeContextBudgetDiagnosticPatches,
+  type ActiveArchivedToolResultPlaceholder,
   type ArchiveRetrievalMode,
   type ContextBudgetPolicy,
   type HistoryCompactBlock,
   type SynthesisSourceRef,
+  type ToolResultArchiveRef,
 } from './context-budget.js';
 import { evaluateHistoryCompactCheckpointReplay } from './history-compact.js';
 import { HistoryCompactSummarizerError } from './history-compact-error.js';
@@ -39,23 +43,56 @@ import {
   type HistoryCompactCheckpoint,
 } from './history-compact-checkpoint.js';
 
+import { createHash } from 'node:crypto';
+import type { ModelMessage } from 'ai';
+import type {
+  ModelAdapter,
+  NormalizedAiSdkUsage,
+  PrepareStepFunctionLike,
+  PrepareStepLike,
+  PrepareStepResultLike,
+} from './model-adapter.js';
+import {
+  activeToolResultLineageIdentity,
+  rewriteActiveToolResultsInMessages,
+  type ActiveToolResultArchiveCandidate,
+  type ActiveToolResultPruneDiagnosticPatch,
+} from './active-tool-result-prune.js';
+import {
+  rewriteActiveFullCompactInMessages,
+  type ActiveCompactionHeadAnchor,
+  type ActiveFullCompactBlock,
+} from './active-full-compact.js';
+import {
+  rewriteSemanticCompactInMessages,
+  type SemanticCompactBlock,
+  type SemanticCompactControllerState,
+} from './semantic-compact.js';
+import { collectStaleToolResultArchiveCandidates } from './tool-result-archive.js';
+
 /** Constructor dependencies for AiSdkCompaction. Grows as more families move in. */
 export interface AiSdkCompactionDeps {
   input: AiSdkBackendInput;
   sessionId: string;
   now: () => number;
+  modelAdapter: ModelAdapter;
+  computeCostUsd: (usage: NormalizedAiSdkUsage) => number | undefined;
 }
 
 export class AiSdkCompaction {
   private readonly input: AiSdkBackendInput;
   private readonly sessionId: string;
   private readonly now: () => number;
+  private readonly modelAdapter: ModelAdapter;
+  private readonly computeCostUsd: (usage: NormalizedAiSdkUsage) => number | undefined;
   private historyCompactAbortController: AbortController | null = null;
 
   constructor(deps: AiSdkCompactionDeps) {
     this.input = deps.input;
     this.sessionId = deps.sessionId;
     this.now = deps.now;
+    this.modelAdapter = deps.modelAdapter;
+    this.computeCostUsd = deps.computeCostUsd;
   }
 
   /** Abort an in-flight manual history compaction (called by AiSdkBackend.stop). */
@@ -772,6 +809,366 @@ export class AiSdkCompaction {
   public hasHistoryCompactCheckpointWriter(): boolean {
     return Boolean(this.input.summarizeHistoryCompact && this.input.recordHistoryCompactCheckpoint);
   }
+
+  public async prepareContextBudgetPolicy(runtimeContext: readonly RuntimeEvent[]): Promise<{
+    policy: ContextBudgetPolicy | undefined;
+    diagnosticPatch?: Partial<ContextBudgetDiagnostic>;
+  }> {
+    const policy = this.input.contextBudget;
+    if (!policy) return { policy };
+    let nextPolicy = policy;
+
+    if (policy.staleToolResultPrune?.enabled === true) {
+      const candidates = collectStaleToolResultArchiveCandidates(
+        runtimeContext,
+        policy?.staleToolResultPrune,
+        policy?.charsPerToken ?? 4,
+        policy?.minRecentTurns,
+      );
+      if (candidates.length > 0) {
+        const archiveRefs = new Map<string, ToolResultArchiveRef>();
+        const existingArchiveRefs = nextPolicy.staleToolResultPrune?.archiveRefs;
+        if (Array.isArray(existingArchiveRefs)) {
+          for (const ref of existingArchiveRefs) archiveRefs.set(ref.runtimeEventId, ref);
+        } else if (existingArchiveRefs) {
+          for (const ref of Object.values(existingArchiveRefs))
+            archiveRefs.set(ref.runtimeEventId, ref);
+        }
+        for (const candidate of candidates) {
+          const bodySha256 = sha256(candidate.serializedResult);
+          const archived = await Promise.resolve(
+            this.input.archiveToolResult?.({
+              ...candidate,
+              sessionId: this.sessionId,
+              bodySha256,
+            }),
+          ).catch(() => undefined);
+          if (!archived?.artifactId) continue;
+          archiveRefs.set(candidate.runtimeEventId, {
+            runtimeEventId: candidate.runtimeEventId,
+            toolCallId: candidate.toolCallId,
+            toolName: candidate.toolName,
+            artifactId: archived.artifactId,
+            bodySha256,
+            originalEstimatedTokens: candidate.originalEstimatedTokens,
+            originalBytes: candidate.originalBytes,
+            rewriteVersion: ARCHIVED_TOOL_RESULT_REWRITE_VERSION,
+            reason: candidate.reason,
+          });
+        }
+
+        nextPolicy = {
+          ...nextPolicy,
+          staleToolResultPrune: {
+            ...nextPolicy.staleToolResultPrune!,
+            archiveRefs: [...archiveRefs.values()],
+          },
+        };
+      }
+    }
+
+    const compactLoadPatch = await this.loadHistoryCompactBlocks(nextPolicy);
+    if (compactLoadPatch.policy !== nextPolicy) nextPolicy = compactLoadPatch.policy;
+    const loadPatch = await this.loadSynthesisCacheBlocks(nextPolicy);
+    if (loadPatch.policy !== nextPolicy) nextPolicy = loadPatch.policy;
+    const diagnosticPatch = mergeContextBudgetDiagnosticPatches(
+      compactLoadPatch.diagnosticPatch,
+      loadPatch.diagnosticPatch,
+    );
+    return {
+      policy: nextPolicy,
+      ...(diagnosticPatch ? { diagnosticPatch } : {}),
+    };
+  }
+
+  public buildActiveToolResultPrunePrepareStep(
+    turnId: string,
+    includeNewestStep: boolean,
+    onDiagnosticPatch?: (patch: ActiveToolResultPruneDiagnosticPatch) => void,
+  ): PrepareStepFunctionLike | undefined {
+    const policy = this.input.contextBudget?.activeToolResultPrune;
+    if (policy?.enabled !== true) return undefined;
+
+    const archivedPlaceholders = new Map<string, ActiveArchivedToolResultPlaceholder>();
+    return async (options) => {
+      const eligibleToolCallIds = collectPrunablePrepareStepToolCallIds(
+        options.steps,
+        includeNewestStep,
+      );
+      if (eligibleToolCallIds.size === 0) return undefined;
+      const rewritten = await rewriteActiveToolResultsInMessages({
+        messages: options.messages,
+        policy,
+        stepNumber: options.stepNumber,
+        turnId,
+        charsPerToken: this.input.contextBudget?.charsPerToken,
+        eligibleToolCallIds,
+        archivedPlaceholders,
+        archiveToolResult: async (candidate) => {
+          return await Promise.resolve(
+            this.input.archiveToolResult?.({
+              ...candidate,
+              sessionId: this.sessionId,
+              runtimeEventId: candidate.runtimeEventId ?? activeToolResultArchiveKey(candidate),
+            }),
+          );
+        },
+      });
+      if (hasActiveToolResultPruneDiagnosticPatch(rewritten.diagnosticPatch)) {
+        onDiagnosticPatch?.(rewritten.diagnosticPatch);
+      }
+      return rewritten.rewritten > 0 ? { messages: rewritten.messages } : undefined;
+    };
+  }
+
+  public buildSemanticCompactPrepareStep(
+    turnId: string,
+    model: unknown,
+    runtimeEvents: readonly RuntimeEvent[] | undefined,
+    headAnchor: ActiveCompactionHeadAnchor | undefined,
+    requestShapeHashForMessages: (
+      messages: readonly ModelMessage[],
+      activeToolsForStep: readonly string[] | undefined,
+    ) => string,
+    onDiagnosticPatch?: (patch: Partial<ContextBudgetDiagnostic>) => void,
+    abortSignal?: AbortSignal,
+  ): PrepareStepFunctionLike | undefined {
+    const policy = this.input.contextBudget?.semanticCompact;
+    if (policy?.enabled !== true || policy.mode === 'off' || !headAnchor) return undefined;
+
+    let acceptedProjection: ActiveFullCompactPrepareStepProjection | undefined;
+    const controllerState: SemanticCompactControllerState = {
+      consecutiveInvalidSummaries: 0,
+      totalInvalidSummaries: 0,
+      compactCallCount: 0,
+      compactCallTotalTokens: 0,
+      acceptedEstimatedTokensSaved: 0,
+    };
+    return async (options) => {
+      const activeToolsForStep = (options as PrepareStepLike & { activeTools?: readonly string[] })
+        .activeTools;
+      const dryRun = policy.mode === 'validate_only' || policy.mode === 'prepare_step_dry_run';
+      const incomingMessages = options.messages;
+      const projectedMessages = dryRun
+        ? undefined
+        : projectAcceptedActiveFullCompactMessages(incomingMessages, acceptedProjection);
+      const messagesForRewrite = projectedMessages ?? incomingMessages;
+      const summarizerModel = policy.summarizerModel
+        ? this.input.modelFactory({
+            connection: this.input.connection,
+            apiKey: this.input.apiKey,
+            modelId: policy.summarizerModel,
+          })
+        : model;
+      const summarizerModelId = policy.summarizerModel ?? this.input.modelId;
+      const rewritten = await rewriteSemanticCompactInMessages({
+        sessionId: this.sessionId,
+        turnId,
+        messages: messagesForRewrite,
+        policy,
+        controllerState,
+        runtimeEvents: runtimeEvents?.filter((event) => event.turnId === turnId),
+        stepNumber: options.stepNumber,
+        now: this.now(),
+        charsPerToken: this.input.contextBudget?.charsPerToken,
+        requestShapeHashForMessages: (messages) =>
+          requestShapeHashForMessages(messages, activeToolsForStep),
+        headAnchor,
+        ...(acceptedProjection?.semanticBlock
+          ? { predecessorBlock: acceptedProjection.semanticBlock }
+          : {}),
+        abortSignal: abortSignal,
+        summarizer: async (request) => {
+          const startedAt = this.now();
+          const callId = `semantic_compact_${turnId}_${options.stepNumber}_${startedAt}`;
+          try {
+            const result = await this.modelAdapter.generateCompactSummary({
+              model: summarizerModel,
+              system: request.system,
+              messages: request.messages,
+              maxOutputTokens: request.maxOutputTokens,
+              abortSignal: request.abortSignal,
+            });
+            this.recordSemanticCompactSummaryCall({
+              callId,
+              turnId,
+              modelId: summarizerModelId,
+              startedAt,
+              latencyMs: Math.max(0, this.now() - startedAt),
+              usage: result.usage,
+              finishReason: result.finishReason,
+              status: 'success',
+            });
+            return result;
+          } catch (error) {
+            this.recordSemanticCompactSummaryCall({
+              callId,
+              turnId,
+              modelId: summarizerModelId,
+              startedAt,
+              latencyMs: Math.max(0, this.now() - startedAt),
+              status: request.abortSignal?.aborted ? 'aborted' : 'error',
+              errorClass: this.modelAdapter.classifyError(error),
+            });
+            throw error;
+          }
+        },
+      });
+      onDiagnosticPatch?.({
+        semanticCompactEnabled: true,
+        semanticCompactMode: policy.mode ?? 'replace',
+        ...rewritten.diagnosticPatch,
+      });
+      if (!dryRun && rewritten.decision === 'replaced') {
+        if (rewritten.block) this.recordSemanticCompactBlock(rewritten.block);
+        acceptedProjection = {
+          sourceSignatures: incomingMessages.map(projectionSourceMessageSignature),
+          sourceSignatureMode: 'active_prune_lineage',
+          projectedMessages: rewritten.messages,
+          ...(rewritten.block ? { semanticBlock: rewritten.block } : {}),
+        };
+        return {
+          messages: rewritten.messages,
+          makaSemanticCompactStatus: 'replaced',
+        } as ActiveCompactionPrepareStepResult;
+      }
+      return !dryRun && projectedMessages
+        ? ({
+            messages: projectedMessages,
+            makaSemanticCompactStatus: 'projected',
+          } as ActiveCompactionPrepareStepResult)
+        : undefined;
+    };
+  }
+
+  public buildActiveFullCompactPrepareStep(
+    turnId: string,
+    runtimeEvents: readonly RuntimeEvent[] | undefined,
+    headAnchor: ActiveCompactionHeadAnchor | undefined,
+    requestShapeHashForMessages: (
+      messages: readonly ModelMessage[],
+      activeToolsForStep: readonly string[] | undefined,
+    ) => string,
+    onDiagnosticPatch?: (patch: Partial<ContextBudgetDiagnostic>) => void,
+  ): PrepareStepFunctionLike | undefined {
+    const policy = this.input.contextBudget?.activeFullCompact;
+    if (policy?.enabled !== true || policy.mode === 'index_only' || policy.mode === 'off')
+      return undefined;
+
+    let acceptedProjection: ActiveFullCompactPrepareStepProjection | undefined;
+    return (options) => {
+      const activeToolsForStep = (options as PrepareStepLike & { activeTools?: readonly string[] })
+        .activeTools;
+      const dryRun = policy.mode === 'validate_only' || policy.mode === 'prepare_step_dry_run';
+      const incomingMessages = options.messages;
+      const projectedMessages = dryRun
+        ? undefined
+        : projectAcceptedActiveFullCompactMessages(incomingMessages, acceptedProjection);
+      const messagesForRewrite = projectedMessages ?? incomingMessages;
+      const rewritten = rewriteActiveFullCompactInMessages({
+        sessionId: this.sessionId,
+        turnId,
+        messages: messagesForRewrite,
+        policy,
+        runtimeEvents: runtimeEvents?.filter((event) => event.turnId === turnId),
+        stepNumber: options.stepNumber,
+        now: this.now(),
+        charsPerToken: this.input.contextBudget?.charsPerToken,
+        requestShapeHashForMessages: (messages) =>
+          requestShapeHashForMessages(messages, activeToolsForStep),
+        ...(headAnchor ? { headAnchor } : {}),
+        dryRun,
+        ...(dryRun ? { dryRunReason: policy.mode } : {}),
+      });
+      onDiagnosticPatch?.(rewritten.diagnosticPatch);
+      if (!dryRun && rewritten.decision === 'replaced') {
+        if (rewritten.block) this.recordActiveFullCompactBlock(rewritten.block);
+        acceptedProjection = {
+          sourceSignatures: incomingMessages.map(modelMessageSignature),
+          sourceSignatureMode: 'exact',
+          projectedMessages: rewritten.messages,
+        };
+        return { messages: rewritten.messages };
+      }
+      return !dryRun && projectedMessages ? { messages: projectedMessages } : undefined;
+    };
+  }
+
+  private recordSemanticCompactSummaryCall(input: {
+    callId: string;
+    turnId: string;
+    modelId: string;
+    startedAt: number;
+    latencyMs: number;
+    usage?: NormalizedAiSdkUsage;
+    finishReason?: string;
+    status: LlmCallRecord['status'];
+    errorClass?: string;
+  }): void {
+    if (!input.usage) return;
+    const costUsd = this.computeCostUsd(input.usage);
+    this.input.recordLlmCall?.({
+      sessionId: this.sessionId,
+      turnId: input.turnId,
+      callKind: 'semantic_compact',
+      callId: input.callId,
+      connectionSlug: this.input.connection.slug,
+      providerId: this.input.connection.providerType,
+      modelId: input.modelId,
+      inputTokens: input.usage.inputTokens,
+      outputTokens: input.usage.outputTokens,
+      cacheHitInputTokens: input.usage.cacheHitInputTokens,
+      cacheMissInputTokens: input.usage.cacheMissInputTokens,
+      ...(input.usage.cacheMissInputSource !== undefined
+        ? { cacheMissInputSource: input.usage.cacheMissInputSource }
+        : {}),
+      cachedInputTokens: input.usage.cachedInputTokens,
+      cacheWriteInputTokens: input.usage.cacheWriteInputTokens,
+      reasoningTokens: input.usage.reasoningTokens,
+      totalTokens: input.usage.totalTokens,
+      ...(input.finishReason !== undefined ? { rawFinishReason: input.finishReason } : {}),
+      ...(input.usage.raw !== undefined ? { rawUsage: input.usage.raw } : {}),
+      latencyMs: input.latencyMs,
+      status: input.status,
+      ...(input.errorClass ? { errorClass: input.errorClass } : {}),
+      startedAt: input.startedAt,
+      ...(costUsd !== undefined ? { costUsd } : {}),
+    });
+  }
+
+  private recordSemanticCompactBlock(block: SemanticCompactBlock): void {
+    const recorder = this.input.recordSemanticCompactBlock;
+    if (!recorder) return;
+    try {
+      const result = recorder(block);
+      if (result && typeof (result as PromiseLike<void>).then === 'function') {
+        void Promise.resolve(result).catch(() => {
+          // Semantic compact persistence is diagnostic/storage-only and must
+          // never perturb provider request projection or tool-loop progress.
+        });
+      }
+    } catch {
+      // Semantic compact persistence is diagnostic/storage-only and must never
+      // perturb provider request projection or tool-loop progress.
+    }
+  }
+
+  private recordActiveFullCompactBlock(block: ActiveFullCompactBlock): void {
+    const recorder = this.input.recordActiveFullCompactBlock;
+    if (!recorder) return;
+    try {
+      const result = recorder(block);
+      if (result && typeof (result as PromiseLike<void>).then === 'function') {
+        void Promise.resolve(result).catch(() => {
+          // Active compact persistence is diagnostic/storage-only and must never
+          // perturb provider request projection or tool-loop progress.
+        });
+      }
+    } catch {
+      // Active compact persistence is diagnostic/storage-only and must never
+      // perturb provider request projection or tool-loop progress.
+    }
+  }
 }
 
 // -- moved helpers (defined in ai-sdk-backend, used only by cache write) -------
@@ -787,4 +1184,160 @@ function mergeCountsInto(
   for (const [key, value] of Object.entries(source ?? {})) {
     target[key] = (target[key] ?? 0) + value;
   }
+}
+
+// -- moved helpers (prepare-step / signature / prune) ------------------------
+
+type ActiveCompactionPrepareStepResult = PrepareStepResultLike & {
+  makaSemanticCompactStatus?: 'replaced' | 'projected';
+};
+
+export function composeActiveCompactionPrepareStep(
+  attention: PrepareStepFunctionLike | undefined,
+  capacity: PrepareStepFunctionLike | undefined,
+): PrepareStepFunctionLike | undefined {
+  if (!attention) return capacity;
+  if (!capacity) return attention;
+  return async (options) => {
+    const attentionResult = (await Promise.resolve(attention(options))) as
+      | ActiveCompactionPrepareStepResult
+      | undefined;
+    if (attentionResult?.makaSemanticCompactStatus === 'replaced') {
+      const { makaSemanticCompactStatus: _status, ...providerResult } = attentionResult;
+      return providerResult;
+    }
+    const capacityResult = await Promise.resolve(
+      capacity({
+        ...options,
+        messages: attentionResult?.messages ?? options.messages,
+        ...(attentionResult?.activeTools ? { activeTools: attentionResult.activeTools } : {}),
+      }),
+    );
+    if (!capacityResult) {
+      if (!attentionResult) return undefined;
+      const { makaSemanticCompactStatus: _status, ...providerResult } = attentionResult;
+      return providerResult;
+    }
+    return {
+      ...attentionResult,
+      ...capacityResult,
+      activeTools: capacityResult.activeTools ?? attentionResult?.activeTools,
+      messages: capacityResult.messages ?? attentionResult?.messages,
+    };
+  };
+}
+
+function activeToolResultArchiveKey(
+  candidate: ActiveToolResultArchiveCandidate & { bodySha256: string },
+): string {
+  return `active:${candidate.turnId}:${candidate.toolCallId}:${candidate.bodySha256}`;
+}
+
+/**
+ * Tool results from the newest completed step have not crossed the provider
+ * boundary yet: prepareStep is invoked immediately before the first request
+ * that could show those results to the model. By default active pruning defers
+ * the newest step and archives only older completed steps, after the model has
+ * had one request in which to consume their exact output.
+ *
+ * `includeNewestStep` widens eligibility to every completed step, including the
+ * newest. The caller sets it when mid-turn capacity compaction is active: the
+ * final-payload verdict may need an oversized newest result pruned to a
+ * placeholder before declaring exhaustion, and capacity/recovery rebuilds
+ * re-materialize raw bodies from the ledger that must be re-archived.
+ */
+function collectPrunablePrepareStepToolCallIds(
+  steps: PrepareStepLike['steps'],
+  includeNewestStep: boolean,
+): Set<string> {
+  const out = new Set<string>();
+  const prunableSteps = includeNewestStep ? steps : steps.slice(0, -1);
+  for (const step of prunableSteps) {
+    for (const call of step.toolCalls ?? []) {
+      if (typeof call.toolCallId === 'string' && call.toolCallId.length > 0) {
+        out.add(call.toolCallId);
+      }
+    }
+  }
+  return out;
+}
+
+export interface ActiveFullCompactPrepareStepProjection {
+  sourceSignatures: readonly string[];
+  sourceSignatureMode: 'exact' | 'active_prune_lineage';
+  projectedMessages: readonly ModelMessage[];
+  semanticBlock?: SemanticCompactBlock;
+}
+
+export function projectAcceptedActiveFullCompactMessages(
+  incomingMessages: readonly ModelMessage[],
+  acceptedProjection: ActiveFullCompactPrepareStepProjection | undefined,
+): ModelMessage[] | undefined {
+  if (!acceptedProjection) return undefined;
+  const sourceSignature =
+    acceptedProjection.sourceSignatureMode === 'active_prune_lineage'
+      ? projectionSourceMessageSignature
+      : modelMessageSignature;
+  if (incomingMessages.length < acceptedProjection.sourceSignatures.length) return undefined;
+  for (let index = 0; index < acceptedProjection.sourceSignatures.length; index += 1) {
+    if (sourceSignature(incomingMessages[index]!) !== acceptedProjection.sourceSignatures[index]) {
+      return undefined;
+    }
+  }
+  return [
+    ...acceptedProjection.projectedMessages,
+    ...incomingMessages.slice(acceptedProjection.sourceSignatures.length),
+  ];
+}
+
+function sha256(text: string): string {
+  return createHash('sha256').update(text).digest('hex');
+}
+
+export function modelMessageSignature(message: ModelMessage): string {
+  return sha256(stableStringifyForSignature(message));
+}
+
+/**
+ * A projection source signature must survive representation-only active
+ * pruning. Preserve every message field except a tool-result payload, whose
+ * raw body and archive placeholder are normalized to the same stable lineage
+ * identity (tool call + original body hash). Any other source mutation still
+ * invalidates the accepted projection.
+ */
+function projectionSourceMessageSignature(message: ModelMessage): string {
+  if (message.role !== 'tool' || !Array.isArray(message.content)) {
+    return modelMessageSignature(message);
+  }
+  const normalizedContent = (message.content as unknown[]).map((part) => {
+    const lineage = activeToolResultLineageIdentity(part);
+    if (!lineage || !part || typeof part !== 'object') return part;
+    const { output: _output, result: _result, ...metadata } = part as Record<string, unknown>;
+    return {
+      ...metadata,
+      makaProjectionToolResultLineage: lineage,
+    };
+  });
+  return modelMessageSignature({ ...message, content: normalizedContent } as ModelMessage);
+}
+
+function stableStringifyForSignature(value: unknown): string {
+  if (value === undefined) return '';
+  if (value === null || typeof value !== 'object') return JSON.stringify(value) ?? '';
+  if (Array.isArray(value)) return `[${value.map(stableStringifyForSignature).join(',')}]`;
+  const object = value as Record<string, unknown>;
+  return `{${Object.keys(object)
+    .sort()
+    .map((key) => `${JSON.stringify(key)}:${stableStringifyForSignature(object[key])}`)
+    .join(',')}}`;
+}
+
+export function hasActiveToolResultPruneDiagnosticPatch(
+  patch: ActiveToolResultPruneDiagnosticPatch,
+): boolean {
+  return (
+    (patch.activePrunedToolResults ?? 0) > 0 ||
+    (patch.activeArchiveFailures ?? 0) > 0 ||
+    (patch.activeEstimatedTokensSaved ?? 0) > 0
+  );
 }
